1.0.3

+++ src/components/or.h
@@ -8,11 +8,11 @@ class OR :
 	public Component
 {
 public:
-	OR(Board* board, Input* inputs, Output* outputs, const unsigned int inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
-	OR(Board* board, Link** inputs, Link** outputs, const unsigned int inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
+	OR(Board* board, Input* inputs, Output* outputs, const size_t inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
+	OR(Board* board, Link** inputs, Link** outputs, const size_t inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
 
 	void compute() override {
-		for (unsigned int i = 0; i < inputCount; i++) {
+		for (size_t i = 0; i < inputCount; i++) {
 			if (inputs[i].getPowered() == true) {
 				outputs[0].setPowered(true);
 				return;
