+++ src/components/xor.h
@@ -8,12 +8,12 @@ class XOR :
 	public Component
 {
 public:
-	XOR(Board* board, Input* inputs, Output* outputs, const unsigned int inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
-	XOR(Board* board, Link** inputs, Link** outputs, const unsigned int inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
+	XOR(Board* board, Input* inputs, Output* outputs, const size_t inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
+	XOR(Board* board, Link** inputs, Link** outputs, const size_t inputCount) : Component(board, inputs, outputs, inputCount, 1) { }
 
 	void compute() override {
-		unsigned int c = 0;
-		for (unsigned int i = 0; i < inputCount; i++) {
+		uint_fast32_t c = 0;
+		for (size_t i = 0; i < inputCount; i++) {
 			c += inputs[i].getPowered();
 		}
 		outputs[0].setPowered(c % 2);
