+++ src/link.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stddef.h>
 
 class Input;
 class Output;
@@ -10,8 +11,8 @@ public:
 	Board* board;
 	Input** inputs;
 	Output** outputs;
-	unsigned int inputCount;
-	unsigned int outputCount;
+	size_t inputCount;
+	size_t outputCount;
 	bool* powered;
 
 	Link() :
@@ -34,7 +35,7 @@ public:
 	{
 	}
 
-	Link(Board* board, Input** inputs, Output** outputs, const unsigned int inputCount, const unsigned int outputCount) :
+	Link(Board* board, Input** inputs, Output** outputs, const size_t inputCount, const size_t outputCount) :
 		board(board),
 		inputs(inputs),
 		outputs(outputs),
